* Refactor the application code so that the UI components can be generic.
