+++ components/src/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 
-import { UserForm } from "./components/UserForm";
+import { DocumentUserForm } from "./components/DocumentUserForm";
 import { Provider } from 'mobx-react'
 import { stores } from "./stores/index";
 import { Component } from "react";
@@ -11,7 +11,7 @@ export default class RootComponent extends Component {
     render() {
       return (
         <Provider {...stores}>
-          <UserForm />
+          <DocumentUserForm />
         </Provider>
       )
     }
